% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figas_install.R
\name{figas_install}
\alias{figas_install}
\title{Set-up environment to run Fine-Grained Aspect-based Sentiment analysis. This is a one-time operation, that should be run when you install the SentiBigNomicsR package, keeping all arguments to their default values.}
\usage{
figas_install(
  conda = "auto",
  version = "latest",
  spacy_version = "2.2.3",
  python_version = "3.7",
  envname = "figas_condaenv",
  pip = FALSE,
  python_path = NULL,
  prompt = FALSE
)
}
\arguments{
\item{conda}{The path to a `conda` executable (argument in `reticulate`). Use `"auto"` to
automatically find an appropriate `conda` binary (default option).}

\item{version}{Version of the `pandas`, `numpy` and `nltk` modules. Use `"latest"` to download the latest available versions (default option).}

\item{spacy_version}{Version of the `spaCy` module to be installed. Default is "2.2.3".}

\item{python_version}{Python version to be used. Default is "3.7".}

\item{envname}{Name of the conda virtual environment. Default is "figas_condaenv".}

\item{pip}{Logical. If `FALSE` (default option), then do not use `pip` when downloading or installing packages (argument in `reticulate`).}

\item{python_path}{Path to the Python installion to be used. If `FALSE`, it automatically finds the Python installation (default option).}

\item{prompt}{Logical. Get prompt messages during the installation. Default to `FALSE`.}
}
\description{
Set-up environment to run Fine-Grained Aspect-based Sentiment analysis. This is a one-time operation, that should be run when you install the SentiBigNomicsR package, keeping all arguments to their default values.
}
\examples{
## Compute the economic sentiment from two texts about two tokens of interest, namely "unemployment" and "economy".
library(SentiBigNomicsR)
figas_install()
text <- list("Unemployment is rising at high speed", "The economy is slowing down and unemployment is booming")
include = list("unemployment", "economy")
my_sent      <- get_sentiment(text = text, include = include)
my_sent
}
\seealso{
\code{\link{get_sentiment}} to compute sentiment.
}
