% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senti_bignomics.R
\docType{data}
\name{senti_bignomics}
\alias{senti_bignomics}
\title{Sentiment of economic terms.}
\format{
A tibble with 3,181 rows and 2 variables:
\describe{
  \item{term}{economic term}
  \item{sentiment}{sentiment score defined over [-1,1]}
}
}
\usage{
data(senti_bignomics)
}
\description{
A dataset containg the sentiment score for 3,181 economic terms.
The score is defined over [-1,1] and was obtained from combining
human annotions from 10 US-based annotators.
}
\references{
Consoli, Sergio, Luca Barbaglia, and Sebastiano Manzan. "Fine-grained, aspect-based sentiment analysis on economic and financial lexicon." Knowledge-Based Systems (2022): 108781. \url{https://doi.org/10.1016/j.knosys.2022.108781}
}
\seealso{
\code{\link{get_sentiment}} to compute the sentiment.
}
\keyword{datasets}
