% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sentiment.R
\name{get_sentiment}
\alias{get_sentiment}
\title{Compute Fine-Grained Aspect-based Sentiment analysis}
\usage{
get_sentiment(
  text,
  include,
  exclude = NULL,
  location = NULL,
  tense = list("past", "present", "future", "NaN"),
  oss = FALSE
)
}
\arguments{
\item{text}{A list of one or more texts.}

\item{include}{A list of tokens of interest about which you want to compute the sentiment.}

\item{exclude}{A list of tokens to exclude when calculating the sentiment.}

\item{location}{A list of locations by which you want to filter your texts.}

\item{tense}{A list of tenses to consider. Allowed values are: \code{"past"}, \code{"present"}, \code{"future"}, \code{"NaN"}. The default contains all these values.}

\item{oss}{Logical. If TRUE, computes the Overall Sentiment Score}
}
\value{
sentiment A tibble "sentiment" containing the average sentiment computed for each text:
 \item{"Doc_id"}{A interger to identify the n^{th} item of the text input}
 \item{"Average_sentiment"}{The average sentiment score defined over [-1,1]}

sentiment_by_chunk A tibble "sentiment_by_chunk" containing the sentiment computed for each chunk detected in the texts.
 \item{"Doc_id"}{A interger to identify the n^{th} item of the text input}
 \item{"Text"}{Text given as input}
 \item{"Chunk"}{Chunk of sentence extracted from Text}
 \item{"Sentiment"}{Sentiment score defined over [-1,1]}
 \item{"Tense"}{Detected tense (possible values are \code{"past"}, \code{"present"}, \code{"future"}, \code{"NaN"})}
 \item{"Include"}{Word which the sentiment is referring to}
}
\description{
\code{get_sentiment} takes in a list of texts and compute the sentiment associated to a list of token of interest.
}
\examples{
## Compute the economic sentiment from two texts about two tokens of interest, namely "unemployment" and "economy".
library(SentiBigNomicsR)
text <- list("Unemployment is rising at high speed", "The economy is slowing down and unemployment is booming")
include = list("unemployment", "economy")
my_sent      <- get_sentiment(text = text, include = include)
my_sent

## Compute the sentiment about "economy", excluding "unemployment"
my_sent      <- get_sentiment(text = list("The economy is slowing down and unemployment is booming"), include = list("economy"), exclude = list("unemployment"))
my_sent

}
\references{
Consoli, Barbaglia, Manzan (January 14, 2021). Fine-Grained  Aspect-Based Sentiment Analysis on Economic and Financial Lexicon. Available at SSRN: https://ssrn.com/abstract=3766194
\url{https://ssrn.com/abstract=3766194}
}
\seealso{
\code{\link{figas_install}}
}
